People Model (Vuex)
- Created new model for people.
- Modified functions to get data from people store model.
- Remove functionality people (characters) from store file.
