+++ src/views/FilmDetail.vue
@@ -48,7 +48,7 @@
         </b-card-group>
       </b-col>
     </b-row>
-    <ModalPeople :info="$store.state.characterDetails" @reset="resetModal" />
+    <ModalPeople :info="$store.state.people.characterDetails" @reset="resetModal" />
     <!-- <p>{{ $store.state.characterDetails }}</p> -->
   </div>
 </template>
